requirements.txt
streamlit
pandas
numpy
scikit-learn
nltk
joblib
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
